Order Details page

+++ source/ApotekaShop.Web/Helpers/CurrencyHelper.cs
@@ -10,7 +10,7 @@ namespace ApotekaShop.Web.Helpers
         public static string Currency(this HtmlHelper helper, int? price, bool onlyPrice = false)
         {
             if (!price.HasValue) return string.Empty;
-            decimal currentPrice = price.Value/100;
+            double currentPrice = price.Value/100.00;
             return onlyPrice ? currentPrice.ToString("F2", CultureInfo.InvariantCulture) : String.Format("{0} {1}", ShopResources.CurrencyName, currentPrice.ToString("F2",CultureInfo.InvariantCulture));
         }
     }
